Move segutil from PS to revcache
The functionality is added functionality to the revocation cache.
It should therefore be in the revcache package so that it can be used by other services as well.
