+++ go/lib/infra/modules/segfetcher/BUILD.bazel
@@ -8,6 +8,7 @@ go_library(
         "request.go",
         "requester.go",
         "resolver.go",
+        "revocation.go",
         "segreplyhandler.go",
         "segs.go",
         "splitter.go",
