+++ go/lib/revcache/BUILD.bazel
@@ -2,13 +2,17 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = ["revcache.go"],
+    srcs = [
+        "revcache.go",
+        "util.go",
+    ],
     importpath = "github.com/scionproto/scion/go/lib/revcache",
     visibility = ["//visibility:public"],
     deps = [
         "//go/lib/addr:go_default_library",
         "//go/lib/common:go_default_library",
         "//go/lib/ctrl/path_mgmt:go_default_library",
+        "//go/lib/ctrl/seg:go_default_library",
         "//go/lib/infra/modules/cleaner:go_default_library",
         "//go/lib/infra/modules/db:go_default_library",
     ],
@@ -16,16 +20,18 @@ go_library(
 
 go_test(
     name = "go_default_test",
-    srcs = ["revcache_test.go"],
+    srcs = ["util_test.go"],
     embed = [":go_default_library"],
     deps = [
         "//go/lib/addr:go_default_library",
         "//go/lib/common:go_default_library",
         "//go/lib/ctrl/path_mgmt:go_default_library",
+        "//go/lib/ctrl/seg:go_default_library",
         "//go/lib/infra:go_default_library",
         "//go/lib/revcache/mock_revcache:go_default_library",
         "//go/lib/util:go_default_library",
         "//go/lib/xtest:go_default_library",
+        "//go/lib/xtest/graph:go_default_library",
         "//go/proto:go_default_library",
         "@com_github_golang_mock//gomock:go_default_library",
         "@com_github_smartystreets_goconvey//convey:go_default_library",
