+++ go/path_srv/internal/handlers/BUILD.bazel
@@ -34,7 +34,6 @@ go_library(
         "//go/lib/snet:go_default_library",
         "//go/lib/snet/addrutil:go_default_library",
         "//go/lib/topology:go_default_library",
-        "//go/path_srv/internal/segutil:go_default_library",
         "//go/proto:go_default_library",
         "@com_github_opentracing_opentracing_go//:go_default_library",
     ],
