--- go/path_srv/internal/segutil/BUILD.bazel
@@ -1,31 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["segutil.go"],
-    importpath = "github.com/scionproto/scion/go/path_srv/internal/segutil",
-    visibility = ["//go/path_srv:__subpackages__"],
-    deps = [
-        "//go/lib/ctrl/path_mgmt:go_default_library",
-        "//go/lib/ctrl/seg:go_default_library",
-        "//go/lib/revcache:go_default_library",
-    ],
-)
-
-go_test(
-    name = "go_default_test",
-    srcs = ["segutil_test.go"],
-    embed = [":go_default_library"],
-    deps = [
-        "//go/lib/common:go_default_library",
-        "//go/lib/ctrl/path_mgmt:go_default_library",
-        "//go/lib/ctrl/seg:go_default_library",
-        "//go/lib/infra:go_default_library",
-        "//go/lib/revcache:go_default_library",
-        "//go/lib/revcache/mock_revcache:go_default_library",
-        "//go/lib/xtest:go_default_library",
-        "//go/lib/xtest/graph:go_default_library",
-        "@com_github_golang_mock//gomock:go_default_library",
-        "@com_github_smartystreets_goconvey//convey:go_default_library",
-    ],
-)
